backend/requirements.txt
fastapi
uvicorn
opensearch-py
python-dotenv